requirements.txt
# Core Dependencies
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# Hugging Face Libraries
diffusers==0.31.0
transformers==4.41.2
accelerate==0.31.0
safetensors==0.4.3
huggingface-hub==0.23.2

# UI
gradio==4.44.1

# Image Processing
pillow==10.4.0
opencv-python==4.8.1.78
matplotlib==3.8.3

# Data Handling
numpy==1.26.4
pandas==2.2.1
datasets==2.18.0

# Metrics
torchmetrics==1.4.0
torch-fidelity==0.3.0
lpips==0.1.4

# Utilities
tqdm==4.66.5
omegaconf==2.3.0
pyyaml==6.0.1
